salva e apaga digital do sensor

+++ Manager.py
@@ -26,20 +26,25 @@ while(True):
         usuario_senha = getpass.getpass("Digita a nova senha do usuário (deixa em branco para não alterar): ")
         usuario = dao.readUser(usuario_id)
         if(input("Digita 's' se quiseres alterar a digital").lower=='s'):
-            usuario_digital = fp.alterarDigital(usuario.impressao_digital) # a impressao_digital pode ser None. a funcao alterarDigital devolve o indice onde a digital foi salva
+            usuario_digital = fp.salvarDigital(usuario.impressao_digital)
         else: usuario_digital = usuario.impressao_digital
         dao.updateUser(usuario_id,usuario_nome,usuario_senha, usuario_digital)
+        print("Usuário", usuario_id, "alterado com sucesso!")
     elif(opcao==3):
         usuario_id = input("Digita o id do novo usuário:")
         usuario_nome = input("Digita o nome do novo usuário:")
         usuario_senha = getpass.getpass("Digita a senha do novo usuário: ")
-        usuario_digital = fp.criarDigital()
+        usuario_digital = fp.salvarDigital()
         dao.createUser(usuario_nome,usuario_id,usuario_senha,usuario_digital)
+        print("Usuário",usuario_id, "salvo com sucesso!")
     elif(opcao==4):
         usuario_id = input("Digita o id do usuário que queres apagar")
         usuario = dao.readUser(usuario_id)
-        fp.apagarDigital(usuario.impressao_digital)
-        dao.deleteUser(usuario_id)
+        if fp.apagarDigital(usuario.impressao_digital):
+            dao.deleteUser(usuario_id)
+            print("Usuário",usuario_id, "apagado com sucesso!")
+        else:
+            print("Não foi possível apagar usuário", usuario_id)
     elif(opcao==5):
         usuario_id = input("Digita o id do usuário que queres ver")
         print(dao.readUser(usuario_id))
